refactor: AUTH_ERROR_CODE, normalizeQueryError를 분리(#118)
- AUTH_ERROR_CODE, normalizeQueryError를 actions/_queryError 로 분리
